towards num replacement of ebi_arithmetic

+++ src/network_simplex_value_type.rs
@@ -1,5 +1,5 @@
-use ebi_arithmetic::{fraction::Fraction, fraction_f64::FractionF64};
-use num::BigInt;
+use ebi_arithmetic::{Fraction, fraction::fraction_f64::FractionF64};
+use malachite::Integer;
 
 pub trait MulWithFloat {
     fn mul_with_float(self, rhs: &f64) -> Self;
@@ -25,7 +25,7 @@ impl MulWithFloat for i128 {
     }
 }
 
-impl MulWithFloat for BigInt {
+impl MulWithFloat for Integer {
     fn mul_with_float(self, _rhs: &f64) -> Self {
         // this should never occur. it is necessary to make network simplex work on both integers and floats
         panic!("Cannot multiply values of different types");
@@ -39,36 +39,36 @@ impl MulWithFloat for FractionF64 {
 }
 
 pub trait ToBigInt {
-    fn to_big_int(&self) -> BigInt;
+    fn to_big_int(&self) -> Integer;
 }
 
 impl ToBigInt for f64 {
     // this should never occur. it is necessary to make network simplex work on both integers and floats
-    fn to_big_int(&self) -> BigInt {
+    fn to_big_int(&self) -> Integer {
         panic!("Cannot multiply values of different types");
     }
 }
 
 impl ToBigInt for i64 {
-    fn to_big_int(&self) -> BigInt {
-        BigInt::from(*self)
+    fn to_big_int(&self) -> Integer {
+        Integer::from(*self)
     }
 }
 
 impl ToBigInt for i128 {
-    fn to_big_int(&self) -> BigInt {
-        BigInt::from(*self)
+    fn to_big_int(&self) -> Integer {
+        Integer::from(*self)
     }
 }
 
-impl ToBigInt for BigInt {
-    fn to_big_int(&self) -> BigInt {
+impl ToBigInt for Integer {
+    fn to_big_int(&self) -> Integer {
         self.clone()
     }
 }
 
 impl ToBigInt for Fraction {
-    fn to_big_int(&self) -> BigInt {
+    fn to_big_int(&self) -> Integer {
         panic!("Cannot multiply values of different types");
     }
 }
